Rename: service 내부 디렉터리 생성 및 이동
