+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/domain/water_place/RescueSupply.java
@@ -1,4 +1,4 @@
-package kr.ac.kumoh.illdang100.tovalley.domain.waterplace;
+package kr.ac.kumoh.illdang100.tovalley.domain.water_place;
 
 import lombok.*;
