+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/service/domain/accident/AccidentService.java
@@ -1,4 +1,4 @@
-package kr.ac.kumoh.illdang100.tovalley.service.domain;
+package kr.ac.kumoh.illdang100.tovalley.service.domain.accident;
 
 import static kr.ac.kumoh.illdang100.tovalley.dto.page.MainPageRespDto.*;
