+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/web/page/MainPageApiController.java
@@ -1,9 +1,8 @@
 package kr.ac.kumoh.illdang100.tovalley.web.page;
 
-import kr.ac.kumoh.illdang100.tovalley.domain.ProvinceEnum;
 import kr.ac.kumoh.illdang100.tovalley.dto.ResponseDto;
-import kr.ac.kumoh.illdang100.tovalley.service.domain.AccidentService;
-import kr.ac.kumoh.illdang100.tovalley.service.domain.WaterPlaceService;
+import kr.ac.kumoh.illdang100.tovalley.service.domain.accident.AccidentService;
+import kr.ac.kumoh.illdang100.tovalley.service.domain.water_place.WaterPlaceService;
 import kr.ac.kumoh.illdang100.tovalley.service.page.MainPageService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
